refactor: move external type imports to api/types
