+++ src/tmdb/apis/people/types.ts
@@ -1,6 +1,6 @@
 import { TMDBResponseList } from 'src/types/TMDBResponseList'
 import { Movie } from '../movies/types'
-import { TVShow } from 'src/types/TVShow'
+import { TVShow } from '../tv/types'
 import { LanguageCode } from 'src/types/LanguageCode'
 import { Nullable } from 'src/types/Nullable'
 import { TMDBResponse } from 'src/types/TMDBResponse'
